Fix, improve and optimize recent experiments
First, the float-type columns were replaced with decimal type,
so they can be part of index to speed up later the queries.
Then the missing part of query for dimension in correlations
experiment was introduced. Finally the distributions were renamed
in overlapping experiment in case there will be more in the future.