Support no common area in overlapping experiment
In rare case there may be no common area between bounding boxes,
which is observed as a negative value after subtracting mins
from maxes – and therefore an exception is raised when calculating
the logarithm value. In such case we set the common volume
to 0 (which is equal to -inf in logarithmic scale).